requirements.txt
APScheduler==3.6.3
charset-normalizer==2.0.12
click==8.0.3
python-dateutil==2.8.2
python-telegram-bot==13.11
requests==2.27.1
sqlparse==0.4.2

